Text wraparound bug came back
I was trying to be more correct about clipping by doing it on a character-by-character basis, as observed in game, but this also brings back the display of "tttttt" etc. in tutorial 5. I think the key is to do clipping per-character but to also avoid advancing if we don't draw a character. This seems to work.