Clip off-screen lines
The game may wrap one coordinate from a line, so in draw.cpp we see that as a very long line that extends past the edge of the screen. This isn't correct.

You can see this when holding an in-room sensor just barely past the left edge of the screen, for example:

http://localhost:8000/#ESi1L/0EMF0JAPLOKCtwiUAB2GGSPDMkAAAAKlIgQAIBiESglBW8gvAih4Ubn3kTR0uT9U/ep84DCY2LaWqrJEk7Ou6Q9UFUBHn7/v9xSVL/B0Zeos7YZAMIeCyFi97G2n5jjP1tv1G8/cQiZFz/raIiFAKBPp9jv97j6XR+o7t7pBj3LjCB9+AkEAybyYbDmUgeVIOSmIhwMIi5YC4ZjSZjqVQsisIojICx1iIWBVMyAAUn9O8AuQf0YYEVAAmeX8osw16IwIGLpWQ54NxNaMsfaEFfmAwK5AduDZIDJz+Q4VphL/WUoFZhL8IBPakwoncZcQG4Djg8uOpyHgt0K9tvSpga1g3gCGIGAOC7KyAB7QKBgFmvgBhoAggJru0CeoBgYApwIQWAfqBeK6AYZDjpIYwNYGiD3g==