Experimenting with making save files from scratch
This "minimal" save file is a lab created from scratch with a hex
editor, intended to be nearly minimal. It won't crash if you turn on the
remote, and you can navigate between rooms. It's already useful for
observing how the engine special-cases various objects and rooms.
The cursor (object 0) starts on room 1, and all other objects are
arrayed into a grid on room 0x3F. The player sprite is a box, and all
other sprites are a hollow rectangle. All rooms have the same layout,
visible corners but with each side open.