fix for paused game immediately after loading a file
This removes code that was intending to save/restore the game speed
around gameMenu and fileManager modals, but because these different
modals can overlap we would sometimes store '0' as the saved speed,
and confusingly resume directly into a paused game. The pause
indication is subtle, and the game otherwise looks hung.