Don't try to autoSave inside TUT7
We technically can save in TUT7 because it runs in LAB.EXE, but there's
no gameplay reason to support it and the resulting URLs are often
too large because we don't include TUT7 in the compression dictionary.

If you still do want to save in TUT7 for whatever reason, you can press
"S" in the ESC menu and the result will be a working ".lsv" save file.