Replace nipplejs with handwritten joystick
I've been hitting a bug similar to this one:
https://github.com/yoannmoinet/nipplejs/issues/222

In that issue, the nipplejs author replies that they're working on a
library rewrite that they hope will address the issue. This seems like a
good time to re-evaluate our choice of library. This game uses very
little of nipplejs' functionality, and a simpler joystick seems like it
would be better here.

This patch removes nipplejs and adds a joystick that aims to supply the
same game functionality. The appearance is slightly different. Uses CSS
animation for smoothing, just like the GameMenu.