joystick deadzone correction refactor
This patch is intended to make joysticks (onscreen and controller API)
easier to use by applying a correction that was being used internally
for mouse emulation.

The virtual mouse had been correcting for the game's extra wide joystick
deadzone in order to position the mouse accurately. This moves it down a
layer into the joystick device. Coordinates accepted by
Engine.setJoystickAxes() should now approximate a uniform [-1,+1] range
with no deadzone.

As a side effect, the virtual mouse now has a harder time discerning
whether a frame with no motion means the mouse is done or that the
virtual joystick is dithering a stick position near zero. To address
this and hopefully make mouse tracking less brittle overall, this
switches tactics and uses a timer to determine when a positioning
event must be cancelled due to collision.