use sbt86 CPU clock for more output delays
This replaces some of the hardcoded delays, as well as the delays
calculated from sound output, with a new type of delay that measures
CPU cycles elapsed since the entry point.

This uses the sbt86 cycle counter that's only enabled for sound
functions and for cutscenes. It won't capture all original delays,
so we do still need other delay sources, but this is a much nicer
way to capture the timing of collections of sound effects like those
used in cutscenes and teleporters.

testing:
- opening cutscene
- ending cutscene
- sound test
- temporary console logging for queue_delay