sound: fix slight pitch change due to new sample rate
The new audio sample rate of 48kHz no longer exactly divides into the
CPU clock frequency, but we still used an integer accumulator to
place audio impulses. This caused a rounding error that made the audio
play fast by about 6 cents, just on the threshold of audibility.

This patch should restore our previous level of pitch accuracy, by
using a floating point accumulator. CPU clocks are still rounded to
the nearest audio sample, and there may be remaining inaccuracies
in the CPU timing model that we use to generate timestamps. I still
want to find a real hardware reference to compare this to.